+++ apps/dsp-app/src/app/main/form-validators/array-length-greater-than-zero-validator.ts
