+++ apps/dsp-app/src/app/project/list/list-item-form/reusable-list-item-form.component.ts
@@ -1,10 +1,9 @@
 import { Component, EventEmitter, Input, OnDestroy, OnInit, Output } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { StringLiteral } from '@dasch-swiss/dsp-js';
 import { MultiLanguages } from '@dasch-swiss/vre/shared/app-string-literal';
 import { Subscription } from 'rxjs';
 import { startWith } from 'rxjs/operators';
-import { atLeastOneStringRequired } from '../../reusable-project-form/at-least-one-string-required.validator';
+import { atLeastOneStringRequired } from '../../../main/form-validators/at-least-one-string-required.validator';
 
 @Component({
   selector: 'app-reusable-list-item-form',
